+++ Source/MediaInfo/Image/File_Dpx.cpp
@@ -322,7 +322,7 @@ bool File_Dpx::Demux_UnpacketizeContainer_Test()
     {
         size_t* File_Buffer_Size_Hint_Pointer=Config->File_Buffer_Size_Hint_Pointer_Get();
         if (File_Buffer_Size_Hint_Pointer)
-            (*File_Buffer_Size_Hint_Pointer)=File_Size;
+            (*File_Buffer_Size_Hint_Pointer)=(size_t)File_Size;
         return false;
     }
